Add support for client-side flow.
When the 'response_code' query param is set to 'token' an access_token
will be included as a hash fragment so that client-side JavaScript can
access it. To support access token generation, the 'enforce_login'
event's next() callback now requires a user_id as the first argument.

If response_code is ommitted, a default value of 'code' is assumed,
indicating server-side flow.